Fix cursor artifact in the middle of the screen
Even if VNC client mouse cursor is disabled and when using applications
such as Steam or VLC a square artifact is visible in the middle of the
screen where the client cursor would otherwise be. Therefore, instead of
disabling the cursor, enable it but move it to the corner of the screen
and hide it using unclutter. This prevents the artifact appearing.